Defined new functional test
